# Variable ordering for exact inference.