--- a/Project.toml
+++ b/Project.toml
@@ -43,7 +43,7 @@ Reexport = "0.2, 1.0"
 Requires = "1.0.1"
 SpecialFunctions = "0.8,0.10,1.0,1.1,1.2,2"
 StatsBase = "0.25,0.26,0.27,0.28,0.29,0.30,0.31,0.32,0.33"
-julia = "1"
+julia = "1.8"
 
 [extras]
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"