I noticed the following comment in the [exact inference implementation](https://github.com/sisl/BayesNets.jl/blob/6b7ff40a3d653eef9d5626fd43fae652a89d3499/src/Inference/exact.jl#L17).
```
# order impacts performance, but we currently have no ordering heuristics
```
This patch adds an ordering heuristic. The effect on performance can be dramatic: for example, if we load [this network](https://github.com/JixTheCat/Economic-BN/blob/c18597c0478a924b6cf5e0808f01cb179b40466f/riverland.xdsl).

```julia-repl
julia> using BayesNets, BenchmarkTools

julia> bn = readxdsl("riverland.xdsl"); 
```

**without ordering**

```julia-repl
julia> @benchmark infer(
           bn,
           [:Economic_Impact, :Electricity_Use],
           evidence=Assignment(:Clay_Content=>1, :Agrochemical_Type=>2))
BenchmarkTools.Trial: 41 samples with 1 evaluation per sample.
 Range (min … max):  118.883 ms … 129.514 ms  ┊ GC (min … max): 0.00% … 4.63%
 Time  (median):     122.314 ms               ┊ GC (median):    2.85%
 Time  (mean ± σ):   123.157 ms ±   2.325 ms  ┊ GC (mean ± σ):  2.64% ± 1.07%

                  █ ▆                                            
  ▄▁▁▁▁▁▁▄▄▁▁▁▁▄▁▇█▇█▇▇▄▄▇▄▇▄▄▁▁▁▁▁▁▄▄▁▁▁▁▄▄▁▁▁▁▄▁▄▁▁▄▁▁▁▁▄▁▁▁▄ ▁
  119 ms           Histogram: frequency by time          130 ms <

 Memory estimate: 55.27 MiB, allocs estimate: 13050.
```

**with ordering**

```julia-repl
julia> @benchmark infer(
           bn,
           [:Economic_Impact, :Electricity_Use],
           evidence=Assignment(:Clay_Content=>1, :Agrochemical_Type=>2))
BenchmarkTools.Trial: 1605 samples with 1 evaluation per sample.
 Range (min … max):  3.040 ms …  13.740 ms  ┊ GC (min … max): 0.00% … 75.38%
 Time  (median):     3.068 ms               ┊ GC (median):    0.00%
 Time  (mean ± σ):   3.111 ms ± 440.326 μs  ┊ GC (mean ± σ):  0.96% ±  4.51%

   ▆█▅                                                         
  ▄████▆▄▄▃▃▂▂▂▂▂▂▂▁▂▂▁▂▁▁▂▁▂▂▁▂▁▁▁▁▁▂▂▁▁▂▁▂▁▂▂▁▂▁▂▁▁▁▁▁▁▂▁▁▂ ▃
  3.04 ms         Histogram: frequency by time        3.57 ms <

 Memory estimate: 558.35 KiB, allocs estimate: 9740.
```