@mykelk 2025-03-09
Thanks for this great contribution! I think `CI.yml` needs to be updated to use actions/cache v4 so that we can run the tests.