@samuelsonric 2025-03-09
checkout seems to have a version 4 in any case: https://github.com/actions/checkout