@samuelsonric 2025-03-09
CliqueTrees should work with Julia 1.8+. Worth noting that 1.10 is the new LTS!