@mykelk 2025-03-09
I think CliqueTrees requires Julia 1.8 and later. I think the Project.toml needs to be updated to reflect this.