@samuelsonric 2025-03-09
Ok: tests pass locally on Julia 1.8. I updated the Project and the CI.